Ising Model Simulation
This file is a simulation of the ferromagnetic Ising model in one dimension. The purpose of this file is to use monte-carlo simulation techniques to calculate numerical values of energy, magnetization, and entropy in a ferromagnet, then compare those results with the analytical solutions. The code works well, but it will take a considerable amount of time to get an accurate numerical entropy.